Improvements. The number of pipelines has been reduced to two. This way it will be more simple to compare results.
