+++ pyriemann_qiskit/utils/filtering.py
@@ -1,5 +1,5 @@
-from sklearn.base import TransformerMixin
-
+from sklearn.base import BaseEstimator, TransformerMixin
+import numpy as np
 
 class NoDimRed(TransformerMixin):
     """No dimensional reduction.
@@ -112,3 +112,21 @@ class NaiveDimRed(TransformerMixin):
         """
         offset = 0 if self.is_even else 1
         return X[:, offset::2]
+
+
+class Vectorizer(BaseEstimator, TransformerMixin):
+    """This is an auxiliary transformer that allows one to vectorize data
+       structures in a pipeline For instance, in the case of an X with dimensions
+       Nt x Nc x Ns, one might be interested in a new data structure with
+       dimensions Nt x (Nc.Ns)
+    """
+    def __init__(self):
+        pass
+    
+    def fit(self, X, y):
+        """fit."""
+        return self
+    
+    def transform(self, X):
+        """transform. """
+        return np.reshape(X, (X.shape[0], -1))
